Backend Service Drive from A to B and all necessary methods in DriveController

+++ Backend_Service/src/main/java/de/htwsaar/backend_service/messages/Command.java
@@ -3,19 +3,20 @@ package de.htwsaar.backend_service.messages;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.WRAPPER_OBJECT)
+
 public abstract class Command {
 
-    private int direction;
+    private Double direction;
 
-    public Command(int direction) {
+    public Command(Double direction) {
         this.direction = direction;
     }
 
-    public int getDirection() {
+    public Double getDirection() {
         return direction;
     }
 
-    public void setDirection(int direction) {
+    public void setDirection(Double direction) {
         this.direction = direction;
     }
 }
