+++ Backend_Service/src/main/java/de/htwsaar/backend_service/messages/DriveCommand.java
@@ -2,7 +2,7 @@ package de.htwsaar.backend_service.messages;
 
 public class DriveCommand extends Command{
 
-    public DriveCommand(int direction) {
+    public DriveCommand(Double direction) {
         super(direction);
     }
