+++ Backend_Service/src/main/java/de/htwsaar/backend_service/messages/TurnCommand.java
@@ -2,7 +2,7 @@ package de.htwsaar.backend_service.messages;
 
 public class TurnCommand extends Command{
 
-    public TurnCommand(int direction) {
+    public TurnCommand(Double direction) {
         super(direction);
     }
 }
